initialising module & gitignore
added .idea to gitignore + created a directory for the first course. Plan is for this directory to have markdown files and diagrams to explain the content. As much as possible, I'll try to incorporate as much content as possible into a big single java app, with pointers from the markdown files in each directory